requirements.txt
# Core workflow
langgraph
pydantic

# HTTP client (NVIDIA API)
requests

# Environment variable loading
python-dotenv